M5

+++ libs/libclock/src/pq.c
@@ -1,12 +1,13 @@
-#include <stdlib.h>
+/*
+ * Prioirty Queue
+ * Glenn McGuire & Cameron Lonsdale
+ */
+
 #include <clock/pq.h>
+#include <stdlib.h>
 
 static void remove_element(priority_queue *pq, uint32_t id);
 
-/* 
- * Initialis the priority queue 
- * @returns pointer to priority queue
- */
 priority_queue *
 init_pq(void)
 {
@@ -27,7 +28,7 @@ init_pq(void)
  * @returns -1 on error, else a positive uid
  */
 int
-pq_push(priority_queue *pq, uint64_t priority, uint32_t delay, timer_callback_t cb, void *data, uint8_t repeat, uint32_t uid)
+pq_push(priority_queue *pq, uint64_t priority, uint64_t delay, timer_callback_t cb, void *data, uint8_t repeat, uint32_t uid)
 {
     /* Sanity checks */
     if (!pq)
